Split settings and Parakeet support code
