+++ Sources/Capture/CLAUDE.md
@@ -10,21 +10,22 @@
 
 ## Key Files
 
-- `ContextCaptureEngine.swift` — Carbon meeting-hotkey registration, hotkey debounce,
-  accessibility-backed physical dictation trigger detection, and routing into dictation or meeting handlers
+- `ContextCaptureEngine.swift` — accessibility-backed physical trigger detection,
+  shortcut debounce, and routing into dictation or meeting handlers
 
 ## Current Hotkey Flow
 
 - The physical dictation trigger routes into `DictationSessionController`
 - Dictation can run as hands-free toggle or push-to-talk, based on `HotkeyPreferences`
 - `PhysicalDictationTriggerPreferences` stores the configurable trigger binding, defaulting to right Option and supporting modifier-only or keyed chords
-- Hotkey id `3` routes meeting toggles through the app-provided meeting closure
+- The configured meeting physical trigger routes meeting toggles through the
+  app-provided meeting closure
 - Rapid press repeats are ignored using `TranscriptedConstants.hotkeyActionDebounceInterval`
 - Accessibility-backed trigger registration failures surface through `hotkeyError` so the menubar can explain why dictation trigger capture is unavailable
 
 ## Guardrails
 
-- Keep the C-level Carbon callback tiny and bounce into `@MainActor` work
+- Keep callback-style routing tiny and bounce into `@MainActor` work
 - Keep meeting routing separate from dictation routing
 - Do not reintroduce screenshot/OCR assumptions here unless that feature
   returns in the same change
@@ -34,7 +35,7 @@
 After changing this directory:
 
 ```bash
-bash build.sh
+bash build.sh --no-open
 bash run-tests.sh
 ```
