+++ Sources/UI/Settings/TranscriptedSettingsView.swift
@@ -4,311 +4,6 @@ import SwiftUI
 import TranscriptedCore
 import UniformTypeIdentifiers
 
-private struct SettingsSidebarSection: Identifiable {
-    let id: String
-    let title: String?
-    let pages: [TranscriptedSettingsPage]
-
-    static let defaultSections = [
-        SettingsSidebarSection(id: "home", title: nil, pages: [.home]),
-        SettingsSidebarSection(id: "setup", title: "Setup", pages: [.general, .storage, .connectAgent]),
-        SettingsSidebarSection(id: "trust", title: "Trust", pages: [.support, .about])
-    ]
-}
-
-private struct SettingsSidebarRow: View {
-    let page: TranscriptedSettingsPage
-    let isSelected: Bool
-
-    @State private var isHovering = false
-
-    var body: some View {
-        Label(page.title, systemImage: page.systemImage)
-            .frame(maxWidth: .infinity, minHeight: 28, alignment: .leading)
-            .padding(.horizontal, 9)
-            .contentShape(Rectangle())
-            .settingsHoverGlow(
-                isActive: isHovering && !isSelected,
-                cornerRadius: 8,
-                fill: Color.primary.opacity(0.032),
-                stroke: Color.accentColor.opacity(0.14),
-                shadow: Color.accentColor.opacity(0.08),
-                shadowRadius: 7
-            )
-            .onHover { isHovering = $0 }
-    }
-}
-
-private struct GeneralSettingsHeader: View {
-    var body: some View {
-        HStack(spacing: 8) {
-            Image(systemName: "gearshape.fill")
-                .font(.system(size: 14, weight: .semibold))
-                .foregroundStyle(Color.primary)
-                .frame(width: 24, height: 24)
-                .background(
-                    RoundedRectangle(cornerRadius: 7, style: .continuous)
-                        .fill(Color.primary.opacity(0.14))
-                )
-
-            Text("General")
-                .font(.system(size: 24, weight: .semibold))
-        }
-    }
-}
-
-private struct GeneralSettingsGroup<Content: View>: View {
-    @ViewBuilder var content: Content
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            content
-        }
-        .frame(maxWidth: 680, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 16, style: .continuous)
-                .fill(Color(nsColor: .controlBackgroundColor).opacity(0.88))
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 16, style: .continuous))
-        .overlay(
-            RoundedRectangle(cornerRadius: 16, style: .continuous)
-                .stroke(Color.primary.opacity(0.11), lineWidth: 1)
-        )
-    }
-}
-
-private struct GeneralInfo {
-    let title: String
-    let message: String
-}
-
-private struct GeneralInfoButton: View {
-    let info: GeneralInfo
-
-    @State private var isPresented = false
-    @State private var isHovering = false
-
-    var body: some View {
-        Button {
-            isPresented.toggle()
-        } label: {
-            Image(systemName: "info.circle.fill")
-                .font(.system(size: 12, weight: .semibold))
-                .foregroundStyle(isHovering ? Color.primary : Color.secondary)
-                .frame(width: 18, height: 18)
-                .background(
-                    Circle()
-                        .fill(Color.primary.opacity(isHovering ? 0.10 : 0.04))
-                )
-                .contentShape(Circle())
-        }
-        .buttonStyle(.plain)
-        .help("Learn about \(info.title)")
-        .accessibilityLabel(Text("About \(info.title)"))
-        .popover(isPresented: $isPresented, arrowEdge: .top) {
-            VStack(alignment: .leading, spacing: 8) {
-                Text(info.title)
-                    .font(.subheadline.weight(.semibold))
-                    .foregroundStyle(Color.primary)
-
-                Text(info.message)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            .padding(14)
-            .frame(width: 260, alignment: .leading)
-        }
-        .onHover { isHovering = $0 }
-    }
-}
-
-private struct GeneralTitleLabel: View {
-    let title: String
-    let info: GeneralInfo?
-
-    var body: some View {
-        HStack(spacing: 5) {
-            Text(title)
-                .font(.subheadline.weight(.semibold))
-                .foregroundStyle(Color.primary)
-                .lineLimit(1)
-                .minimumScaleFactor(0.82)
-
-            if let info {
-                GeneralInfoButton(info: info)
-            }
-        }
-        .layoutPriority(1)
-    }
-}
-
-private struct GeneralSectionHeading: View {
-    let title: String
-    let info: GeneralInfo?
-
-    var body: some View {
-        HStack(spacing: 5) {
-            Text(title)
-                .font(.subheadline.weight(.semibold))
-                .foregroundStyle(.secondary)
-
-            if let info {
-                GeneralInfoButton(info: info)
-            }
-        }
-        .padding(.leading, 10)
-    }
-}
-
-private struct GeneralToggleRow: View {
-    let title: String
-    @Binding var isOn: Bool
-    var help: String
-    var info: GeneralInfo? = nil
-
-    var body: some View {
-        HStack(spacing: 10) {
-            GeneralTitleLabel(title: title, info: info)
-
-            Spacer(minLength: 10)
-
-            Toggle(title, isOn: $isOn)
-                .labelsHidden()
-                .toggleStyle(.switch)
-                .controlSize(.regular)
-                .tint(.accentColor)
-                .help(help)
-                .accessibilityLabel(Text(title))
-                .accessibilityValue(Text(isOn ? "On" : "Off"))
-                .accessibilityHint(Text(help))
-        }
-        .padding(.horizontal, 14)
-        .frame(minHeight: 44)
-        .overlay(alignment: .bottom) {
-            Divider()
-        }
-    }
-}
-
-private struct GeneralActionRow: View {
-    let title: String
-    let value: String
-    let systemImage: String?
-    let help: String
-    let action: () -> Void
-
-    @State private var isHovering = false
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: 10) {
-                Text(title)
-                    .font(.subheadline.weight(.semibold))
-                    .foregroundStyle(Color.primary)
-
-                Spacer(minLength: 10)
-
-                HStack(spacing: 6) {
-                    if let systemImage {
-                        Image(systemName: systemImage)
-                            .font(.system(size: 11, weight: .semibold))
-                    }
-                    Text(value)
-                        .font(.caption.weight(.semibold))
-                    Image(systemName: "chevron.right")
-                        .font(.system(size: 9, weight: .bold))
-                        .foregroundStyle(.tertiary)
-                }
-                .foregroundStyle(Color.secondary)
-            }
-            .padding(.horizontal, 14)
-            .frame(minHeight: 44)
-            .frame(maxWidth: .infinity)
-            .contentShape(Rectangle())
-            .background(isHovering ? Color.primary.opacity(0.035) : Color.clear)
-        }
-        .buttonStyle(.plain)
-        .help(help)
-        .onHover { isHovering = $0 }
-        .overlay(alignment: .bottom) {
-            Divider()
-        }
-        .accessibilityLabel(Text(title))
-        .accessibilityValue(Text(value))
-        .accessibilityHint(Text(help))
-    }
-}
-
-private struct GeneralDisclosureRow: View {
-    let title: String
-    let value: String
-    @Binding var isExpanded: Bool
-    let help: String
-    let action: () -> Void
-
-    @State private var isHovering = false
-
-    var body: some View {
-        Button {
-            withAnimation(.snappy(duration: 0.18)) {
-                isExpanded.toggle()
-            }
-            action()
-        } label: {
-            HStack(spacing: 10) {
-                Text(title)
-                    .font(.subheadline.weight(.semibold))
-                    .foregroundStyle(Color.primary)
-
-                Spacer(minLength: 10)
-
-                Text(value)
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-
-                Image(systemName: "chevron.down")
-                    .font(.system(size: 9, weight: .bold))
-                    .foregroundStyle(.secondary)
-                    .frame(width: 20, height: 20)
-                    .background(
-                        Circle()
-                            .fill(Color.primary.opacity(isHovering ? 0.10 : 0.06))
-                    )
-                    .rotationEffect(.degrees(isExpanded ? 180 : 0))
-            }
-            .padding(.horizontal, 14)
-            .frame(minHeight: 44)
-            .frame(maxWidth: .infinity)
-            .contentShape(Rectangle())
-            .background(isHovering ? Color.primary.opacity(0.035) : Color.clear)
-        }
-        .buttonStyle(.plain)
-        .help(help)
-        .onHover { isHovering = $0 }
-        .accessibilityLabel(Text(title))
-        .accessibilityValue(Text("\(value), \(isExpanded ? "expanded" : "collapsed")"))
-        .accessibilityHint(Text(help))
-    }
-}
-
-private struct GeneralExpandedContent<Content: View>: View {
-    @ViewBuilder var content: Content
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            content
-        }
-        .padding(12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(Color.primary.opacity(0.025))
-        .overlay(alignment: .top) {
-            Divider()
-        }
-    }
-}
-
 struct TranscriptedSettingsView: View {
     @Bindable var navigation: TranscriptedSettingsNavigationModel
     @ObservedObject var speakerPeopleModel: SpeakerPeopleSettingsViewModel
@@ -3659,1000 +3354,4 @@ struct TranscriptedSettingsView: View {
         formatter.timeStyle = .short
         return formatter
     }()
-}
-
-private struct AutoEnterAppCandidate: Identifiable, Equatable {
-    let bundleID: String
-    let name: String
-
-    var id: String { bundleID }
-
-    static func runningApps() -> [AutoEnterAppCandidate] {
-        let transcriptedBundleID = Bundle.main.bundleIdentifier
-        let candidates = NSWorkspace.shared.runningApplications.compactMap { app -> AutoEnterAppCandidate? in
-            guard app.activationPolicy == .regular,
-                  let bundleID = app.bundleIdentifier,
-                  bundleID != transcriptedBundleID else {
-                return nil
-            }
-
-            return AutoEnterAppCandidate(
-                bundleID: bundleID,
-                name: app.localizedName ?? bundleID
-            )
-        }
-
-        var seen = Set<String>()
-        return candidates
-            .filter { seen.insert($0.bundleID).inserted }
-            .sorted { lhs, rhs in
-                lhs.name.localizedCaseInsensitiveCompare(rhs.name) == .orderedAscending
-            }
-    }
-}
-
-private struct CorrectionDraftRow: Identifiable, Equatable {
-    let id: UUID
-    var spoken: String
-    var replacement: String
-
-    init(id: UUID = UUID(), spoken: String = "", replacement: String = "") {
-        self.id = id
-        self.spoken = spoken
-        self.replacement = replacement
-    }
-
-    init(entry: CustomDictionaryEntry) {
-        self.init(spoken: entry.spoken, replacement: entry.replacement)
-    }
-
-    static func rows(from rawText: String) -> [CorrectionDraftRow] {
-        let rows = CustomDictionaryPreferences.entries(from: rawText).map(CorrectionDraftRow.init(entry:))
-        return rows.isEmpty ? [CorrectionDraftRow()] : rows
-    }
-
-    static func rawText(from rows: [CorrectionDraftRow]) -> String {
-        rows.compactMap { row in
-            let spoken = row.spoken.trimmingCharacters(in: .whitespacesAndNewlines)
-            let replacement = row.replacement.trimmingCharacters(in: .whitespacesAndNewlines)
-
-            guard !spoken.isEmpty else { return nil }
-            if replacement.isEmpty || replacement == spoken {
-                return spoken
-            }
-            return "\(spoken) -> \(replacement)"
-        }
-        .joined(separator: "\n")
-    }
-}
-
-private struct CorrectionEditorRow: View {
-    @Binding var spoken: String
-    @Binding var replacement: String
-    let onRemove: () -> Void
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 12) {
-            TextField("okay ours", text: $spoken)
-                .textFieldStyle(.roundedBorder)
-                .accessibilityLabel(Text("Mistake"))
-
-            TextField("OKRs", text: $replacement)
-                .textFieldStyle(.roundedBorder)
-                .accessibilityLabel(Text("Fix"))
-
-            Button(role: .destructive, action: onRemove) {
-                Image(systemName: "minus.circle.fill")
-                    .foregroundStyle(.secondary)
-                    .frame(width: 28, height: 28)
-            }
-            .buttonStyle(SettingsHoverButtonStyle(tone: .destructive, cornerRadius: 7))
-            .accessibilityLabel(Text("Remove correction"))
-            .help("Remove this correction.")
-        }
-    }
-}
-
-private struct ModelChoiceRow: View {
-    let model: TranscriptionModelChoice
-    let isPreferred: Bool
-    let isEffective: Bool
-
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            Image(systemName: symbolName)
-                .font(.system(size: 15, weight: .semibold))
-                .foregroundStyle(symbolColor)
-                .frame(width: 22)
-                .padding(.top, 2)
-
-            VStack(alignment: .leading, spacing: 4) {
-                HStack(spacing: 8) {
-                    Text(model.title)
-                        .font(.subheadline.weight(.semibold))
-
-                    Text(statusLabel)
-                        .font(.caption2.weight(.semibold))
-                        .foregroundStyle(statusColor)
-                }
-
-                Text(model.summary)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-
-            Spacer(minLength: 12)
-        }
-    }
-
-    private var symbolName: String {
-        if isEffective { return "checkmark.circle.fill" }
-        return "circle"
-    }
-
-    private var symbolColor: Color {
-        if isEffective { return .green }
-        return .secondary
-    }
-
-    private var statusLabel: String {
-        if isEffective { return "Active" }
-        if isPreferred { return "Preferred" }
-        return model.availabilityStatus
-    }
-
-    private var statusColor: Color {
-        if isEffective { return .green }
-        return .secondary
-    }
-}
-
-private struct AutoEnterAllowedAppRow: View {
-    let title: String
-    let bundleID: String
-    let remove: () -> Void
-
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            Image(systemName: "checkmark.circle.fill")
-                .font(.system(size: 15, weight: .semibold))
-                .foregroundStyle(.green)
-                .padding(.top, 2)
-
-            VStack(alignment: .leading, spacing: 2) {
-                Text(title)
-                    .font(.subheadline.weight(.semibold))
-
-                Text(bundleID)
-                    .font(.caption2)
-                    .foregroundStyle(.secondary)
-            }
-
-            Spacer(minLength: 12)
-
-            SettingsInlineActionButton(
-                title: "Remove",
-                tone: .destructive,
-                action: remove
-            )
-        }
-    }
-}
-
-private struct SettingsRecentMeetingAudioControl: View {
-    let title: String
-    let symbolName: String
-    let isActive: Bool
-    let isPlaying: Bool
-    let scrubber: AnyView?
-    let action: () -> Void
-
-    init(
-        title: String,
-        symbolName: String,
-        isActive: Bool,
-        isPlaying: Bool,
-        scrubber: AnyView? = nil,
-        action: @escaping () -> Void
-    ) {
-        self.title = title
-        self.symbolName = symbolName
-        self.isActive = isActive
-        self.isPlaying = isPlaying
-        self.scrubber = scrubber
-        self.action = action
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Button(action: action) {
-                HStack(spacing: 8) {
-                    Image(systemName: symbolName)
-                        .font(.system(size: 9, weight: .bold))
-                        .foregroundStyle(iconForeground)
-                        .frame(width: 20, height: 20)
-                        .background(Circle().fill(iconBackground))
-
-                    ZStack(alignment: .leading) {
-                        Capsule()
-                            .fill(Color.secondary.opacity(0.22))
-                            .frame(width: 44, height: 4)
-
-                        Capsule()
-                            .fill(playheadColor)
-                            .frame(width: isPlaying ? 28 : 8, height: 4)
-
-                        Circle()
-                            .fill(playheadColor)
-                            .frame(width: 8, height: 8)
-                            .offset(x: isPlaying ? 24 : 4)
-                    }
-                    .frame(width: 44, height: 12)
-
-                    Text(title)
-                        .font(.caption.weight(.semibold))
-                        .lineLimit(1)
-                }
-                .padding(.horizontal, 10)
-                .padding(.vertical, 6)
-            }
-            .buttonStyle(SettingsHoverButtonStyle(
-                tone: isActive ? .accent : .neutral,
-                cornerRadius: 8,
-                normalFill: background,
-                normalStroke: stroke
-            ))
-
-            if let scrubber {
-                scrubber
-            }
-        }
-    }
-
-    private var background: Color {
-        isActive ? Color.accentColor.opacity(0.14) : Color.secondary.opacity(0.08)
-    }
-
-    private var stroke: Color {
-        isActive ? Color.accentColor.opacity(0.28) : Color.primary.opacity(0.10)
-    }
-
-    private var iconBackground: Color {
-        isActive ? Color.accentColor : Color.primary.opacity(0.12)
-    }
-
-    private var iconForeground: Color {
-        isActive ? .white : .secondary
-    }
-
-    private var playheadColor: Color {
-        isActive ? .accentColor : .secondary.opacity(0.65)
-    }
-}
-
-private struct SettingsFailedMeetingRow: View {
-    let item: MeetingSessionController.FailedMeetingItem
-    let canRetry: Bool
-    let retryUnavailableReason: String?
-    let audio: MeetingAudioAttachment?
-    let retryAction: () -> Void
-    let revealAudioAction: () -> Void
-    let secondaryAction: () -> Void
-    @ObservedObject private var playback = MeetingAudioPlayback.shared
-    @State private var selectedPlaybackChoiceID: String?
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 12) {
-            if item.failureKind == .recordingTooShort {
-                Image(systemName: "timer")
-                    .font(.system(size: 14, weight: .semibold))
-                    .foregroundStyle(.secondary)
-                    .frame(width: 24)
-            } else {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .font(.system(size: 14, weight: .semibold))
-                    .foregroundStyle(.orange)
-                    .frame(width: 24)
-            }
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(item.title)
-                    .font(.subheadline.weight(.semibold))
-
-                Text(item.detail)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-
-                Text(item.meta)
-                    .font(.caption2)
-                    .foregroundStyle(.tertiary)
-            }
-
-            Spacer(minLength: 12)
-
-            if let audio {
-                let selectedChoice = selectedPlaybackChoice(for: audio)
-                SettingsRecentMeetingAudioControl(
-                    title: playback.buttonTitle(for: audio, choice: selectedChoice),
-                    symbolName: playback.symbolName(for: audio, choice: selectedChoice),
-                    isActive: playback.isActive(audio, choice: selectedChoice),
-                    isPlaying: playback.isPlaying && playback.isActive(audio, choice: selectedChoice),
-                    scrubber: playback.isActive(audio)
-                        ? AnyView(MeetingAudioScrubber(attachment: audio, width: 190))
-                        : nil
-                ) {
-                    playback.toggle(audio, choice: selectedChoice)
-                }
-                .help("\(playback.buttonTitle(for: audio, choice: selectedChoice)) retained meeting audio")
-
-                MeetingAudioSourceMenu(
-                    attachment: audio,
-                    selectedChoiceID: selectedPlaybackChoiceBinding(for: audio)
-                ) { choice in
-                    if playback.isActive(audio) {
-                        playback.switchSource(audio, choice: choice)
-                    }
-                }
-
-                Button {
-                    revealAudioAction()
-                } label: {
-                    Label("Show Audio", systemImage: "folder")
-                        .font(.caption.weight(.semibold))
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 6)
-                }
-                .buttonStyle(SettingsHoverButtonStyle(
-                    cornerRadius: 8,
-                    normalFill: Color.primary.opacity(0.025),
-                    normalStroke: Color.primary.opacity(0.06)
-                ))
-            }
-
-            if item.isRetryable || item.isRetrying {
-                Button {
-                    retryAction()
-                } label: {
-                    Label(item.isRetrying ? "Retrying..." : "Try Again", systemImage: "arrow.clockwise")
-                        .font(.caption.weight(.semibold))
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 6)
-                }
-                .buttonStyle(SettingsHoverButtonStyle(
-                    tone: .accent,
-                    cornerRadius: 8,
-                    normalFill: Color.accentColor.opacity(0.08),
-                    normalStroke: Color.accentColor.opacity(0.16)
-                ))
-                .disabled(retryDisabled)
-                .help(retryHelp)
-            }
-
-            Button(role: hasRetainedAudioFiles ? .destructive : nil) {
-                secondaryAction()
-            } label: {
-                Label(hasRetainedAudioFiles ? "Delete" : "Dismiss", systemImage: hasRetainedAudioFiles ? "trash" : "xmark")
-                    .font(.caption.weight(.semibold))
-                    .padding(.horizontal, 10)
-                    .padding(.vertical, 6)
-            }
-            .buttonStyle(SettingsHoverButtonStyle(
-                tone: hasRetainedAudioFiles ? .destructive : .neutral,
-                cornerRadius: 8,
-                normalFill: hasRetainedAudioFiles ? Color.red.opacity(0.06) : Color.primary.opacity(0.025),
-                normalStroke: hasRetainedAudioFiles ? Color.red.opacity(0.14) : Color.primary.opacity(0.06)
-            ))
-        }
-    }
-
-    private var retryDisabled: Bool {
-        !canRetry || !item.isRetryable || !item.hasAudioFiles || item.isRetrying
-    }
-
-    private var retryHelp: String {
-        if item.isRetrying {
-            return "Retry is already running."
-        }
-        if !item.hasAudioFiles {
-            return "This meeting does not have enough saved audio to retry."
-        }
-        if !item.isRetryable {
-            return "This meeting does not have enough saved audio to retry."
-        }
-        if let retryUnavailableReason {
-            return retryUnavailableReason
-        }
-        if !canRetry {
-            return "Wait for the current meeting work to finish before retrying."
-        }
-        return "Transcribe this saved audio again."
-    }
-
-    private var hasRetainedAudioFiles: Bool {
-        !item.audioURLs.isEmpty
-    }
-
-    private func selectedPlaybackChoice(for audio: MeetingAudioAttachment) -> MeetingAudioPlaybackChoice? {
-        playback.activeChoice(for: audio) ?? audio.playbackChoice(id: selectedPlaybackChoiceID)
-    }
-
-    private func selectedPlaybackChoiceBinding(for audio: MeetingAudioAttachment) -> Binding<String?> {
-        Binding(
-            get: { selectedPlaybackChoice(for: audio)?.id },
-            set: { selectedPlaybackChoiceID = $0 }
-        )
-    }
-}
-
-private struct AgentConnectionSettingsPage: View {
-    @StateObject private var viewModel = AgentConnectionViewModel(
-        context: AgentConnectionContext(meetingTitle: nil, meetingDate: nil, transcriptURL: nil)
-    )
-    @State private var claudeDesktopStatus = ClaudeDesktopIntegrationInstaller.currentStatus()
-    @State private var claudeDesktopInstallResult: ClaudeDesktopIntegrationInstallResult?
-    @State private var claudeDesktopInstallError: String?
-    @State private var isInstallingClaudeDesktop = false
-    @State private var copiedClaudeDesktopConfig = false
-    @State private var copiedLocalAgentPrompt = false
-    @State private var copiedFolderPrompt = false
-    @State private var copiedFolderPaths = false
-    @State private var showAdvancedAgentSetup = false
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 24) {
-            SettingsPageIntro(
-                title: "Agent",
-                summary: "Install direct tools for Claude Desktop, or copy a prompt for local coding agents."
-            )
-
-            agentActionSection
-
-            if let claudeDesktopInstallResult {
-                ClaudeDesktopSelfTestResultView(result: claudeDesktopInstallResult)
-            }
-
-            if let claudeDesktopInstallError {
-                Label(claudeDesktopInstallError, systemImage: "exclamationmark.triangle.fill")
-                    .font(.caption)
-                    .foregroundStyle(.orange)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-
-            SettingsSection(
-                title: "Details",
-                detail: "Advanced setup."
-            ) {
-                AgentSetupDetailsDisclosure(isExpanded: $showAdvancedAgentSetup) {
-                    VStack(alignment: .leading, spacing: 14) {
-                        ClaudeDesktopStatusRow(status: claudeDesktopStatus)
-
-                        HStack(spacing: 10) {
-                            SettingsInlineActionButton(
-                                title: copiedClaudeDesktopConfig ? "Copied" : "Copy Claude Config",
-                                symbolName: "doc.on.doc"
-                            ) {
-                                copyText(
-                                    ClaudeDesktopIntegrationInstaller.configSnippet(),
-                                    showingCopiedFeedback: $copiedClaudeDesktopConfig
-                                )
-                            }
-
-                            SettingsInlineActionButton(title: "Show Config", symbolName: "folder") {
-                                revealClaudeDesktopConfig()
-                            }
-                            .disabled(!claudeDesktopStatus.configExists)
-                        }
-
-                        AgentFolderRow(
-                            name: "Meetings",
-                            detail: "Meeting Markdown files.",
-                            path: viewModel.context.meetingsFolderURL.path,
-                            isAvailable: viewModel.fileExists(viewModel.context.meetingsFolderURL)
-                        ) {
-                            viewModel.reveal(viewModel.context.meetingsFolderURL)
-                        }
-
-                        AgentFolderRow(
-                            name: "Dictation",
-                            detail: "Dictation Markdown files.",
-                            path: viewModel.context.dictationsFolderURL.path,
-                            isAvailable: viewModel.fileExists(viewModel.context.dictationsFolderURL)
-                        ) {
-                            viewModel.reveal(viewModel.context.dictationsFolderURL)
-                        }
-
-                        Divider()
-
-                        VStack(alignment: .leading, spacing: 8) {
-                            Label("Web chats are fallback only", systemImage: "globe")
-                                .font(.subheadline.weight(.semibold))
-
-                            Text("Claude web, ChatGPT web, Cowork, and mobile chats usually cannot see your Mac. Use them for a pasted meeting or granted folders, not full Transcripted memory.")
-                                .font(.caption)
-                                .foregroundStyle(.secondary)
-                                .fixedSize(horizontal: false, vertical: true)
-
-                            HStack(spacing: 10) {
-                                SettingsInlineActionButton(
-                                    title: copiedFolderPrompt ? "Copied" : "Copy Folder Prompt",
-                                    symbolName: "doc.on.doc"
-                                ) {
-                                    copyText(
-                                        AgentConnectionGuide.folderAccessPrompt,
-                                        showingCopiedFeedback: $copiedFolderPrompt
-                                    )
-                                }
-
-                                SettingsInlineActionButton(
-                                    title: copiedFolderPaths ? "Copied" : "Copy Paths",
-                                    symbolName: "folder"
-                                ) {
-                                    copyText(
-                                        AgentConnectionGuide.folderPathsText,
-                                        showingCopiedFeedback: $copiedFolderPaths
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        .onAppear(perform: refreshClaudeDesktopStatus)
-    }
-
-    private var agentActionSection: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack(spacing: 12) {
-                AgentConnectActionButton(
-                    symbolName: claudeDesktopActionSymbol,
-                    title: claudeDesktopActionTitle,
-                    subtitle: "Claude Desktop",
-                    statusText: claudeDesktopStatusText,
-                    statusSymbolName: claudeDesktopStatusSymbol,
-                    tint: claudeDesktopStatusTint,
-                    isEnabled: claudeDesktopActionEnabled
-                ) {
-                    installClaudeDesktop()
-                }
-
-                AgentConnectActionButton(
-                    symbolName: copiedLocalAgentPrompt ? "checkmark" : "chevron.left.forwardslash.chevron.right",
-                    title: copiedLocalAgentPrompt ? "Copied" : "Copy for Agent",
-                    subtitle: "Codex, Claude Code, Cursor",
-                    statusText: "Local files",
-                    statusSymbolName: "folder",
-                    tint: Color(nsColor: .systemBlue),
-                    isEnabled: true
-                ) {
-                    copyText(
-                        AgentConnectionGuide.starterPrompt(filename: nil),
-                        showingCopiedFeedback: $copiedLocalAgentPrompt
-                    )
-                }
-            }
-
-            if !claudeDesktopStatus.claudeDesktopLikelyInstalled {
-                SettingsInlineActionButton(title: "Get Claude Desktop", symbolName: "arrow.down.circle", tone: .accent) {
-                    openClaudeDesktopDownload()
-                }
-            }
-        }
-    }
-
-    private var claudeDesktopActionTitle: String {
-        if isInstallingClaudeDesktop {
-            return "Installing..."
-        }
-
-        switch claudeDesktopStatus.state {
-        case .installed:
-            return "Install in Claude"
-        case .notInstalled:
-            return "Install in Claude"
-        case .needsRepair:
-            return "Repair Claude Setup"
-        }
-    }
-
-    private var claudeDesktopActionSymbol: String {
-        if isInstallingClaudeDesktop {
-            return "hourglass"
-        }
-
-        switch claudeDesktopStatus.state {
-        case .installed:
-            return "checkmark"
-        case .notInstalled:
-            return "sparkles"
-        case .needsRepair:
-            return "arrow.clockwise"
-        }
-    }
-
-    private var claudeDesktopActionEnabled: Bool {
-        !isInstallingClaudeDesktop
-            && claudeDesktopStatus.bundledBinaryExists
-    }
-
-    private var claudeDesktopStatusText: String {
-        if !claudeDesktopStatus.bundledBinaryExists {
-            return "Missing"
-        }
-
-        switch claudeDesktopStatus.state {
-        case .installed:
-            return "Installed"
-        case .notInstalled:
-            return "Not installed"
-        case .needsRepair:
-            return "Repair"
-        }
-    }
-
-    private var claudeDesktopStatusSymbol: String {
-        switch claudeDesktopStatus.state {
-        case .installed:
-            return "checkmark.circle.fill"
-        case .notInstalled:
-            return "arrow.right.circle"
-        case .needsRepair:
-            return "exclamationmark.triangle.fill"
-        }
-    }
-
-    private var claudeDesktopStatusTint: Color {
-        switch claudeDesktopStatus.state {
-        case .installed:
-            return .green
-        case .notInstalled:
-            return Color(nsColor: .systemOrange)
-        case .needsRepair:
-            return .orange
-        }
-    }
-
-    private func refreshClaudeDesktopStatus() {
-        claudeDesktopStatus = ClaudeDesktopIntegrationInstaller.currentStatus()
-    }
-
-    private func installClaudeDesktop() {
-        guard !isInstallingClaudeDesktop else { return }
-        isInstallingClaudeDesktop = true
-        claudeDesktopInstallResult = nil
-        claudeDesktopInstallError = nil
-
-        Task {
-            do {
-                let result = try await Task.detached(priority: .userInitiated) {
-                    try ClaudeDesktopIntegrationInstaller.installForClaudeDesktop()
-                }.value
-
-                claudeDesktopInstallResult = result
-                refreshClaudeDesktopStatus()
-            } catch {
-                claudeDesktopInstallError = error.localizedDescription
-                refreshClaudeDesktopStatus()
-            }
-
-            isInstallingClaudeDesktop = false
-        }
-    }
-
-    private func copyText(_ text: String, showingCopiedFeedback copiedFlag: Binding<Bool>? = nil) {
-        let pasteboard = NSPasteboard.general
-        pasteboard.clearContents()
-        pasteboard.setString(text, forType: .string)
-
-        guard let copiedFlag else { return }
-        copiedFlag.wrappedValue = true
-        Task { @MainActor in
-            try? await Task.sleep(nanoseconds: 1_500_000_000)
-            copiedFlag.wrappedValue = false
-        }
-    }
-
-    private func revealClaudeDesktopConfig() {
-        NSWorkspace.shared.activateFileViewerSelecting([claudeDesktopStatus.configURL])
-    }
-
-    private func openClaudeDesktopDownload() {
-        guard let url = URL(string: "https://claude.ai/download") else { return }
-        NSWorkspace.shared.open(url)
-    }
-}
-
-private struct AgentConnectActionButton: View {
-    let symbolName: String
-    let title: String
-    let subtitle: String
-    let statusText: String
-    let statusSymbolName: String
-    let tint: Color
-    let isEnabled: Bool
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            HStack(alignment: .top, spacing: 14) {
-                Image(systemName: symbolName)
-                    .font(.system(size: 18, weight: .semibold))
-                    .foregroundStyle(tint)
-                    .frame(width: 42, height: 42)
-                    .background(tint.opacity(0.13), in: RoundedRectangle(cornerRadius: 8, style: .continuous))
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 8, style: .continuous)
-                            .stroke(tint.opacity(0.18), lineWidth: 1)
-                    )
-
-                VStack(alignment: .leading, spacing: 8) {
-                    Text(title)
-                        .font(.title3.weight(.semibold))
-                        .foregroundStyle(.primary)
-                        .lineLimit(1)
-                        .minimumScaleFactor(0.88)
-
-                    Text(subtitle)
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                        .lineLimit(1)
-                        .minimumScaleFactor(0.9)
-
-                    Label(statusText, systemImage: statusSymbolName)
-                        .font(.caption.weight(.semibold))
-                        .foregroundStyle(tint)
-                        .lineLimit(1)
-                        .minimumScaleFactor(0.88)
-                        .padding(.top, 2)
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-                Image(systemName: "arrow.right")
-                    .font(.system(size: 12, weight: .bold))
-                    .foregroundStyle(tint.opacity(isEnabled ? 0.95 : 0.45))
-                    .frame(width: 26, height: 26)
-                    .background(tint.opacity(isEnabled ? 0.11 : 0.06), in: RoundedRectangle(cornerRadius: 7, style: .continuous))
-            }
-            .padding(16)
-            .frame(maxWidth: .infinity, minHeight: 112, alignment: .leading)
-            .contentShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
-        }
-        .buttonStyle(SettingsHoverButtonStyle(
-            tone: .accent,
-            cornerRadius: 8,
-            normalFill: Color(nsColor: .controlBackgroundColor).opacity(0.78),
-            normalStroke: tint.opacity(0.28),
-            hoverFill: Color(nsColor: .controlBackgroundColor).opacity(0.95),
-            pressedFill: Color(nsColor: .controlBackgroundColor).opacity(0.88),
-            hoverStroke: tint.opacity(0.56)
-        ))
-        .disabled(!isEnabled)
-    }
-}
-
-private struct AgentSetupDetailsDisclosure<Content: View>: View {
-    @Binding var isExpanded: Bool
-    let content: Content
-
-    init(
-        isExpanded: Binding<Bool>,
-        @ViewBuilder content: () -> Content
-    ) {
-        self._isExpanded = isExpanded
-        self.content = content()
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Button {
-                withAnimation(.snappy(duration: 0.18)) {
-                    isExpanded.toggle()
-                }
-            } label: {
-                HStack(spacing: 10) {
-                    Image(systemName: "chevron.right")
-                        .font(.system(size: 11, weight: .bold))
-                        .foregroundStyle(.secondary)
-                        .rotationEffect(.degrees(isExpanded ? 90 : 0))
-                        .frame(width: 16, height: 16)
-
-                    Text("Show setup details")
-                        .font(.subheadline.weight(.semibold))
-                        .foregroundStyle(.primary)
-
-                    Spacer(minLength: 12)
-
-                    Text(isExpanded ? "Hide" : "Show")
-                        .font(.caption.weight(.semibold))
-                        .foregroundStyle(.secondary)
-                }
-                .padding(.horizontal, 12)
-                .padding(.vertical, 10)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .contentShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
-            }
-            .buttonStyle(SettingsHoverButtonStyle(
-                tone: .neutral,
-                cornerRadius: 8,
-                normalFill: Color.primary.opacity(0.018),
-                normalStroke: Color.primary.opacity(0.07),
-                hoverFill: Color.primary.opacity(0.04),
-                pressedFill: Color.primary.opacity(0.06),
-                hoverStroke: Color.primary.opacity(0.12)
-            ))
-            .accessibilityLabel(Text("Show setup details"))
-            .accessibilityValue(Text(isExpanded ? "Expanded" : "Collapsed"))
-            .accessibilityHint(Text(isExpanded ? "Hide advanced agent setup details" : "Show advanced agent setup details"))
-
-            if isExpanded {
-                VStack(alignment: .leading, spacing: 14) {
-                    content
-                }
-                .padding(.top, 14)
-                .transition(.opacity.combined(with: .move(edge: .top)))
-            }
-        }
-    }
-}
-
-private struct ClaudeDesktopStatusRow: View {
-    let status: ClaudeDesktopIntegrationStatus
-
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            Image(systemName: symbolName)
-                .font(.system(size: 16, weight: .semibold))
-                .foregroundStyle(tint)
-                .frame(width: 28, height: 28)
-                .background(tint.opacity(0.12), in: RoundedRectangle(cornerRadius: 8, style: .continuous))
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(.subheadline.weight(.semibold))
-
-                Text(detail)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-
-                if let configuredPath = status.configuredCommandPath,
-                   configuredPath != status.installedBinaryURL.path {
-                    Text(configuredPath)
-                        .font(.system(.caption2, design: .monospaced))
-                        .foregroundStyle(.tertiary)
-                        .lineLimit(2)
-                        .textSelection(.enabled)
-                }
-            }
-
-            Spacer(minLength: 12)
-        }
-    }
-
-    private var symbolName: String {
-        switch status.state {
-        case .installed:
-            return "checkmark.circle.fill"
-        case .notInstalled:
-            return "circle"
-        case .needsRepair:
-            return "exclamationmark.triangle.fill"
-        }
-    }
-
-    private var tint: Color {
-        switch status.state {
-        case .installed:
-            return .green
-        case .notInstalled:
-            return .secondary
-        case .needsRepair:
-            return .orange
-        }
-    }
-
-    private var title: String {
-        switch status.state {
-        case .installed:
-            return "Installed"
-        case .notInstalled:
-            return "Not installed yet"
-        case .needsRepair:
-            return "Needs update"
-        }
-    }
-
-    private var detail: String {
-        if !status.bundledBinaryExists {
-            return "This app build does not include Transcripted direct tools yet."
-        }
-
-        if !status.configIsReadable {
-            return "Claude Desktop config is not readable JSON. Install will back it up and write a clean config."
-        }
-
-        switch status.state {
-        case .installed:
-            return "Claude Desktop is configured. Restart Claude Desktop if you just installed it."
-        case .notInstalled:
-            return status.claudeDesktopLikelyInstalled
-                ? "Click Install for Claude Desktop, then restart Claude Desktop."
-                : "Claude Desktop was not found. You can still install now, then install Claude Desktop."
-        case .needsRepair:
-            if !status.installedBinaryExists {
-                return "The server file is missing. Install will copy a fresh one and update Claude Desktop."
-            }
-            if !status.installedBinaryMatchesBundled {
-                return "The installed server is from an older app build. Install will copy the current one."
-            }
-            return "Claude Desktop points at another Transcripted server. Install will update it."
-        }
-    }
-}
-
-private struct ClaudeDesktopSelfTestResultView: View {
-    let result: ClaudeDesktopIntegrationInstallResult
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Label(
-                "Ready. Restart Claude.",
-                systemImage: "checkmark.circle.fill"
-            )
-            .foregroundStyle(.green)
-
-            Text("\(result.selfTest.meetingFileCount) meetings, \(result.selfTest.dictationFileCount) dictation files found.")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-
-            if let backupURL = result.backupURL {
-                Text("Previous config backed up to \(backupURL.lastPathComponent).")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .font(.caption)
-        .fixedSize(horizontal: false, vertical: true)
-    }
-}
-
-private struct AgentFolderRow: View {
-    let name: String
-    let detail: String
-    let path: String
-    let isAvailable: Bool
-    let action: () -> Void
-
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            VStack(alignment: .leading, spacing: 4) {
-                HStack(spacing: 8) {
-                    Text(name)
-                        .font(.subheadline.weight(.semibold))
-
-                    if !isAvailable {
-                        Text("Not written yet")
-                            .font(.caption2.weight(.semibold))
-                            .foregroundStyle(.orange)
-                    }
-                }
-
-                Text(detail)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-
-                Text(path)
-                    .font(.system(.caption2, design: .monospaced))
-                    .foregroundStyle(.secondary)
-                    .textSelection(.enabled)
-            }
-
-            Spacer(minLength: 12)
-
-            SettingsInlineActionButton(title: "Reveal", action: action)
-                .disabled(!isAvailable)
-        }
-    }
 }
